move minion collector into a separate mod
