+++ src/game/minion/mod.rs
@@ -2,7 +2,9 @@ use bevy::prelude::*;
 use bevy_rapier3d::prelude::Collider;
 
 mod walk_target;
+mod collector;
 
+pub use collector::*;
 pub use walk_target::*;
 
 use super::{CharacterWalkControl, KinematicCharacterBundle, PlayerTag};
